refactor: :fire: fixed import issues and added things back that were removed while resolving conflict
